refactor(luna-stage): consolidate lynx-stage into `./lynx` directory and mark `web-core` as optional peer dep
Move `src/components/lynx-stage/` into `src/lynx/` so the directory name
directly corresponds to the `./lynx` package entry point, mirroring the
existing `src/motion/` → `./motion` convention. Declare `@lynx-js/web-core`
as an optional peer dependency since it is only required when using the
`./lynx` entry point.

- Move `LynxStage`, `LunaLynxStage`, `useLynxStage` and related files into `src/lynx/`
- Remove `src/components/lynx-stage/` directory
- Update internal import paths accordingly
- Add `@lynx-js/web-core >=0.20.0` to `peerDependencies` with `optional: true`
- No API or behavior changes
