+++ apps/studio/src/components/lynx-stage/index.ts
@@ -2,6 +2,6 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-export { LynxStage } from '@dugyu/luna-stage';
+export { LynxStage } from '@dugyu/luna-stage/lynx';
 export { StudioLunaLynxStage } from './studio-luna-lynx-stage';
 export type { StudioLunaLynxStageProps } from './studio-luna-lynx-stage';
