+++ packages/luna-stage/src/lynx/index.ts
@@ -3,10 +3,8 @@
 // LICENSE file in the root directory of this source tree.
 
 export { LynxStage } from './lynx-stage';
-export type { LynxStageProps } from './lynx-stage';
-
-export { useLynxStage } from './use-lynx-stage';
-export type { UseLynxStageOptions, UseLynxStageResult } from './use-lynx-stage';
-
 export { LunaLynxStage } from './luna-lynx-stage';
+export { useLynxStage } from './use-lynx-stage';
+export type { LynxStageProps } from './lynx-stage';
 export type { LunaLynxStageProps } from './luna-lynx-stage';
+export type { UseLynxStageOptions, UseLynxStageResult } from './use-lynx-stage';
