+++ packages/luna-stage/src/lynx/luna-lynx-stage.tsx
@@ -14,8 +14,8 @@ import {
 } from './lynx-stage-constants';
 import { useLynxStage } from './use-lynx-stage';
 import type { UseLynxStageOptions } from './use-lynx-stage';
-import { useIsClient } from '../../hooks';
-import type { LynxGlobalProps } from '../../types/lynx-view';
+import { useIsClient } from '../hooks';
+import type { LynxGlobalProps } from '../types/lynx-view';
 
 export type LunaLynxStageProps =
   & Omit<UseLynxStageOptions, 'globalProps' | 'bundleBaseUrl'>
